% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneId.R
\docType{data}
\name{modelGeneId}
\alias{modelGeneId}
\title{IdentifiHR: gene annotations and beta coefficients}
\format{
A data frame with 2604 rows and 3 columns:
\describe{
  \item{ENSEMBL}{A character vector of ensembl identifiers}
  \item{betaCoeff}{A numeric vector of weighted genes in the trained IdentifiHR model}
  \item{hgnc_symbol}{A character vector of hgnc symbols}
  \item{entrezgene_id}{A numeric vector of entrez identifiers}
}
}
\source{
Created for subsetting genes down to those required by identifiHR. Generated from top table of DE analysis.
}
\usage{
modelGeneId
}
\description{
Identifiers for model genes
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
\keyword{datasets}
