% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelIdentifiHR.R
\docType{data}
\name{modelIdentifiHR}
\alias{modelIdentifiHR}
\title{IdentifiHR: trained elastic net penalised regression model}
\format{
An object of class \code{"glmnet"}.
}
\source{
Trained IdentifiHR model for predicting HR status.
}
\usage{
modelIdentifiHR
}
\description{
Trained IdentifiHR logistic regression model that can predict HR status using the weighted expression of 209 genes.
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
\keyword{datasets}
