% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeanGenesIdentifiHR.R
\docType{data}
\name{modelMeanGenesIdentifiHR}
\alias{modelMeanGenesIdentifiHR}
\title{IdentifiHR: means of 2604 DE genes}
\format{
A data frame with 2604 rows and 1 column, with ensembl identifiers given as rownames:
\describe{
  \item{mean}{A numeric vector of mean log2 counts-per-million values for genes required by the identifiHR model}
}
}
\source{
Created from training cohort for z score scaling.
}
\usage{
modelMeanGenesIdentifiHR
}
\description{
Mean expression values for model genes, to be used in z score scaling.
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
\keyword{datasets}
