% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissingness.R
\name{plotMissingness}
\alias{plotMissingness}
\title{Plot which model genes are missing from input data}
\usage{
plotMissingness(y)
}
\arguments{
\item{y}{A data frame of all genes required by IdentifiHR, as outputted by the "interrogateMissingness()" function.}
}
\value{
A plot of the beta coefficients of the genes that contribute to prediction in IdentifiHR, coloured by whether they were "missing" or "present" in the input data.
}
\description{
The "plotMissingness()" function allows you to visualise which weighted genes, if any, required for the IdentifiHR model are missing in your dataset. There are 209 gene required to predict HR status by IdentifiHR, and if any are missing, the model's accuracy will be reduced.
}
\examples{
data(rawCounts)
missingGenes <- interrogateMissingness(y = rawCounts, geneIds = "ENSEMBL")
plotMissingness(missingGenes)
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
