% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProb.R
\name{plotProb}
\alias{plotProb}
\title{Plot the probability that a sample is HRD, as predicted by IdentifiHR}
\usage{
plotProb(y)
}
\arguments{
\item{y}{A data frame outputted by the "predictHr()" function, including columns "Sample", "hrPrediction" and "predictionProbability".}
}
\value{
A plot of the input cohort and the TCGA test cohort, against the probability of each sample being HRD.
}
\description{
The "plotProb()" function allows you to visualise the probability of the input sample/s being "HRD", and to compare this against the probability distribution of the TCGA HGSC testing cohort analysed in our manuscript.
}
\examples{
plotProb(predictions)
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
