% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictHr.R
\name{predictHr}
\alias{predictHr}
\title{Predict HR status using processed gene expresssion counts}
\usage{
predictHr(y)
}
\arguments{
\item{y}{A numeric matrix of gene expression abundances from the output of the "processCounts()" function, with samples presented in rows and genes presented in columns.}
}
\value{
A data frame containing the sample identifier, the predicted HR status and the probability that a given sample if HRP.
\itemize{
  \item Sample - A character of sample identifiers taken from the input.
  \item hrPrediction - A character vector indicating the discrete homologous recombination (HR) status of a sample, being deficient ("HRD") or proficient ("HRP").
  \item predictionProbability - A numeric vector of the probability that a given sample is HRD.
}
}
\description{
The "predictHr" function utilises IdentifiHR, a pre-trained penalised logistic regression classifier, to predict homologous recombination status. While IdentifiHR was trained to classify high-grade serous ovarian carcinomas, it can be applied to other cancer types, though the accuracy of classification will be reduced.
}
\examples{
#See GitHub wiki
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
