% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCounts.R, R/processCountsOld.R
\name{processCounts}
\alias{processCounts}
\title{IdentifiHR}
\usage{
processCounts(y, logCpm, scaled, geneIds)

processCounts(y, logCpm, scaled, geneIds)
}
\arguments{
\item{y}{A numeric matrix of gene expression counts, with genes presented in rows and samples presented in columns.}

\item{logCpm}{Logical. Have the raw gene expression counts been transformed into log2-counts-per-million? Default is TRUE, being that the counts matrix has been transformed. If FALSE, a log2-counts-per-million transformation will be applied.}

\item{scaled}{Logical. Has z-score scaling been performed? Default is TRUE, being that the counts matrix has been scaled. If FALSE, z-score scaling of each gene will be performed.}

\item{geneIds}{How are genes annotated? Specify either "ENSEMBL", "SYMBOL" or "ENTREZID".}
}
\value{
A numeric matrix of counts for model genes, that has been log2-counts-per-million transformed and z-score scaled across genes.

A numeric matrix of counts for model genes, that has been log2-counts-per-million transformed and z-score scaled across genes.
}
\description{
The "processCounts" function prepares gene expression counts for input into the IdentifiHR classifier.

The "processCounts" function prepares gene expression counts for input into the IdentifiHR classifier.
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
