% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaTestPred.R
\docType{data}
\name{tcgaTestPred}
\alias{tcgaTestPred}
\title{IdentifiHR: TCGA HGSC testing cohort HR status predictions}
\format{
A gene expression counts table with 52125 rows, containing ensembl gene identifiers as rownames, and 10 columns, representing distinct samples.
 \describe{
  \item{Sample}{A character of sample identifiers taken from the TCGA HGSC tetsing cohort}
  \item{hrPrediction}{A character vector indicating the discrete HR status of a sample, being deficient ("HRD") or proficient ("HRP")}
  \item{predictionProbability}{A numeric vector of the probability that a given sample is HRD}
}
}
\source{
Created during the testing and validation of the IdentifiHR model.
}
\usage{
tcgaTestPred
}
\description{
HR status predictions in the TCGA HGSC testing cohort.
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
\keyword{datasets}
