% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogateMissingness.R
\name{interrogateMissingness}
\alias{interrogateMissingness}
\title{Interrogate which model genes are missing from input data}
\usage{
interrogateMissingness(y, geneIds)
}
\arguments{
\item{y}{A numeric matrix or data frame of raw gene expression counts, with genes presented as rownames and samples presented in columns.}

\item{geneIds}{How are genes annotated? Specify either "ENSEMBL", "HGNC" or "ENTREZ" ("ENSEMBL" is preferred).}
}
\value{
A data frame containing the following columns:
\itemize{
  \item ensembl_id - A character vector of ensembl identifiers for genes needed in the model.
  \item hgnc_symbol - A character vector of hgnc symbols for genes needed in the model.
  \item entrezgene_id - A numeric vector of entrez identifiers for genes needed in the model.
  \item betaCoef - A numeric vector of beta coefficients for genes that are weighted towards IdentifiHR predictions. NAs indicate that a gene is only used in normalisation.
  \item inputStatus - A character vector indicating if a gene was "present"  or "missing in the input data. 
  \item normWeightedGene - A character vector indicating if a gene contributes to only normalisation, being "normGene", or contributes to both normalisation and model predictions, being "normWeightedGene".
}
}
\description{
The "interrogateMissingness()" function allows you to examine which genes, if any, required for the IdentifiHR model are missing in your dataset. There are 2604 genes required for normalising library size differences and 209 gene required to predict HR status by IdentifiHR.
}
\examples{
data(rawCounts)
# Removing a model gene gene to give an example of the interrogateMissingness() function.
rawCountsMissing <- rawCounts[rownames(rawCounts) != "ENSG00000160959", ]
missingGenes <- interrogateMissingness(y = rawCountsMissing, geneIds = "ENSEMBL")
head(missingGenes)
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
