% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictHr.R
\name{predictHr}
\alias{predictHr}
\title{Predict HR status using processed gene expresssion counts}
\usage{
predictHr(y)
}
\arguments{
\item{y}{A numeric matrix of gene expression abundances from the output of the "processCounts()" function, with samples presented in rows and genes presented in columns.}
}
\value{
A data frame containing the sample identifier, the predicted HR status and the probability that a given sample if HRP.
}
\description{
The "predictHr" function utilises IdentifiHR, a pre-trained penalised logistic regression classifier, to predict homologous recombination status. While IdentifiHR was trained to classify high-grade serous ovarian carcinomas, it can be applied to other cancer types, though the accuracy of classification will be reduced.
}
\examples{
# to add
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
