% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCounts.R
\name{processCounts}
\alias{processCounts}
\title{IdentifiHR}
\usage{
processCounts(y, geneIds)
}
\arguments{
\item{y}{A numeric matrix of raw gene expression counts, with genes presented in rownames and samples presented in columns.}

\item{geneIds}{How are genes annotated? Specify either "ENSEMBL", "HGNC" or "ENTREZ" ("ENSEMBL" is preferred).}
}
\value{
A numeric matrix of counts for model genes, that has been log2-counts-per-million transformed and z-score scaled across genes.
}
\description{
The "processCounts" function prepares gene expression counts for input into the IdentifiHR classifier.
}
\examples{
# to add
}
\author{
Ashley L Weir, \email{weir.a@wehi.edu.au}
}
